# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/properties/extract": {
      "post": {
        "operationId": "PropertiesController_extract",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      }
    },
    "/properties": {
      "post": {
        "operationId": "PropertiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      },
      "get": {
        "operationId": "PropertiesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      }
    },
    "/properties/{id}": {
      "get": {
        "operationId": "PropertiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      },
      "patch": {
        "operationId": "PropertiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      },
      "delete": {
        "operationId": "PropertiesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      }
    },
    "/reports/stats": {
      "get": {
        "operationId": "ReportsController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get aggregated portfolio statistics",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/ai-summary": {
      "post": {
        "operationId": "ReportsController_generateAiSummary",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Generate AI executive summary of the portfolio",
        "tags": [
          "reports"
        ]
      }
    },
    "/storage/upload": {
      "post": {
        "operationId": "StorageController_uploadFile",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Storage"
        ]
      }
    },
    "/tenants": {
      "post": {
        "operationId": "TenantsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      },
      "get": {
        "operationId": "TenantsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      }
    },
    "/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      },
      "patch": {
        "operationId": "TenantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      },
      "delete": {
        "operationId": "TenantsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenants"
        ]
      }
    },
    "/leases": {
      "post": {
        "operationId": "LeasesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Leases"
        ]
      },
      "get": {
        "operationId": "LeasesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Leases"
        ]
      }
    },
    "/leases/{id}": {
      "get": {
        "operationId": "LeasesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Leases"
        ]
      },
      "patch": {
        "operationId": "LeasesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Leases"
        ]
      },
      "delete": {
        "operationId": "LeasesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Leases"
        ]
      }
    },
    "/finances/stats": {
      "get": {
        "operationId": "FinancesController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finances"
        ]
      }
    },
    "/finances": {
      "get": {
        "operationId": "FinancesController_findAll",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finances"
        ]
      },
      "post": {
        "operationId": "FinancesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Finances"
        ]
      }
    },
    "/finances/{id}": {
      "delete": {
        "operationId": "FinancesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finances"
        ]
      }
    },
    "/documents/{propertyId}": {
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/{documentId}/versions": {
      "post": {
        "operationId": "DocumentsController_addVersion",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/{id}": {
      "delete": {
        "operationId": "DocumentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    }
  },
  "info": {
    "title": "Buena Property API",
    "description": "API for Property Dashboard & Creation Flow",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreatePropertyDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}